[ALPS07919787] eint: mt6989: modify data type of pin_list
Original data type of pin_list is not sufficient for
global eint num over 255.

MTK-Commit-Id: 15e03540775f7e924c38cdfee69c3a84bf7238cb

CR-Id: ALPS07919787
Feature: eint

+++ drivers/pinctrl/mediatek/mtk-eint.h
@@ -57,7 +57,7 @@ struct mtk_eint_instance {
 	const char *name;
 	void __iomem *base;
 	unsigned int number;
-	u8 pin_list[MAX_PIN];
+	u16 pin_list[MAX_PIN];
 	unsigned int *wake_mask;
 	unsigned int *cur_mask;
 };
